make wave router composition only
